# Conflicts:
#	core/src/test/java/fr/abes/licencesnationales/core/converter/UtilsMapperTest.java
#	core/src/test/java/fr/abes/licencesnationales/core/services/EtablissementServiceTest.java
#	web/src/main/java/fr/abes/licencesnationales/web/controllers/EtablissementController.java
#	web/src/test/java/fr/abes/licencesnationales/web/controllers/EtablissementControllerTest.java

+++ core/src/main/java/fr/abes/licencesnationales/core/repository/contactediteur/ContactEditeurRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ContactEditeurRepository extends JpaRepository<ContactEditeurEntity, Long> {
+public interface ContactEditeurRepository extends JpaRepository<ContactEditeurEntity, Integer> {
 }
